feat(task): Extract list item in to component
